+++ src/storage/points-reader.h
@@ -31,10 +31,10 @@
 
 typedef struct points_reader_s {
   data_point_t *points;
-  int points_count;
+  uint64_t points_count;
 } points_reader_t;
 
-points_reader_t *points_reader_create(data_point_t *points, int points_count);
+points_reader_t *points_reader_create(data_point_t *points, uint64_t points_count);
 void points_reader_destroy(points_reader_t *reader);
 
 #endif  // SRC_STORAGE_POINTS_READER_H_
